[java] run example files separately from tests
Set examples to run sequentially instead of in parallel with other tests
