+++ java/main/pom.xml
@@ -125,17 +125,41 @@
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-surefire-plugin</artifactId>
         <version>3.1.2</version>
-        <configuration>
-          <properties>
-            <configurationParameters>
-              junit.jupiter.execution.parallel.enabled = true
-              junit.jupiter.execution.parallel.mode.default = concurrent
-              junit.jupiter.execution.parallel.config.strategy = fixed
-              junit.jupiter.execution.parallel.config.fixed.parallelism = ${surefire.parallel}
-              junit.jupiter.execution.parallel.config.fixed.max-pool-size = ${surefire.parallel}
-            </configurationParameters>
-          </properties>
-        </configuration>
+        <executions>
+          <execution>
+            <id>parallel-tests</id>
+            <goals>
+              <goal>test</goal>
+            </goals>
+            <configuration>
+              <properties>
+                <configurationParameters>
+                  junit.jupiter.execution.parallel.enabled = true
+                  junit.jupiter.execution.parallel.mode.default = concurrent
+                  junit.jupiter.execution.parallel.config.strategy = fixed
+                  junit.jupiter.execution.parallel.config.fixed.parallelism = ${surefire.parallel}
+                  junit.jupiter.execution.parallel.config.fixed.max-pool-size = ${surefire.parallel}
+                </configurationParameters>
+              </properties>
+            </configuration>
+          </execution>
+          <execution>
+            <id>sequential-tests</id>
+            <goals>
+              <goal>test</goal>
+            </goals>
+            <configuration>
+              <includes>
+                <include>**/*Example.java</include>
+              </includes>
+              <properties>
+                <configurationParameters>
+                  junit.jupiter.execution.parallel.enabled = false
+                </configurationParameters>
+              </properties>
+            </configuration>
+          </execution>
+        </executions>
       </plugin>
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
