+++ java/main/src/test/java/com/saucelabs/saucebindings/examples/DataCenterExample.java
@@ -4,20 +4,23 @@ import com.saucelabs.saucebindings.*;
 import org.junit.jupiter.api.Test;
 import org.openqa.selenium.remote.RemoteWebDriver;
 
-public class SessionTest {
+public class DataCenterExample {
 
   @Test
-  public void startSession() {
+  public void changeDataCenter() {
     // 1. Create Session object with the defaults
     SauceSession session = new SauceSession();
 
-    // 2. Start Session to get the Driver
+    // 2. Set Data Center
+    session.setDataCenter(DataCenter.EU_CENTRAL);
+
+    // 3. Start Session to get the Driver
     RemoteWebDriver driver = session.start();
 
-    // 3. Use the driver in your tests just like normal
+    // 4. Use the driver in your tests just like normal
     driver.get("https://www.saucedemo.com/");
 
-    // 4. Stop the Session with whether the test passed or failed
+    // 5. Stop the Session with whether the test passed or failed
     session.stop(true);
   }
 }
