+++ java/main/src/test/java/com/saucelabs/saucebindings/examples/DisableExample.java
@@ -5,13 +5,17 @@ import java.util.List;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.parallel.Execution;
+import org.junit.jupiter.api.parallel.ExecutionMode;
 import org.openqa.selenium.WebDriver;
 
-public class DisableTest {
+@Execution(ExecutionMode.SAME_THREAD)
+public class DisableExample {
 
   @Test
   public void startSession() {
     // 1. Toggle off sauce labs
+    System.out.println("disabling sauce");
     System.setProperty("sauce.disabled", "true");
 
     // 2. Create a Sauce Session
@@ -37,5 +41,6 @@ public class DisableTest {
   @AfterEach
   public void stopSession() {
     System.clearProperty("sauce.disabled");
+    System.out.println("clearing disable");
   }
 }
