+++ java/main/src/test/java/com/saucelabs/saucebindings/examples/SauceLabsOptionsExample.java
@@ -2,19 +2,20 @@ package com.saucelabs.saucebindings.examples;
 
 import com.saucelabs.saucebindings.*;
 import com.saucelabs.saucebindings.options.SauceOptions;
+import java.time.Duration;
 import org.junit.jupiter.api.Test;
 import org.openqa.selenium.remote.RemoteWebDriver;
 
-public class CommonOptionsTest {
+public class SauceLabsOptionsExample {
 
   @Test
-  public void basicOptions() {
-    // 1. Create SauceOptions instance with common w3c options
+  public void sauceOptions() {
+    // 1. Specify Sauce Specific Options Based on Browser
     SauceOptions sauceOptions =
         SauceOptions.firefox()
-            .setBrowserVersion("127.0")
-            .setPlatformName(SaucePlatform.WINDOWS_10)
-            .setUnhandledPromptBehavior(UnhandledPromptBehavior.IGNORE)
+            .setExtendedDebugging()
+            .setIdleTimeout(Duration.ofSeconds(45))
+            .setTimeZone("Alaska")
             .build();
 
     // 2. Create Session object with the Options object instance
